Archived forms
* archived form components into components/forms
* removed the forms page, couldn't figure out how to re-route so went
  with the easier option of just removing the page from our domain
