+++ components/forms/SubmittedForms.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Header from "./Header";
+import Header from "../Header";
 import SubmittedFormsArray from "./SubmittedFormsArray";
 import { Col, Row } from "react-bootstrap";
 import Form from "./SubmittedForm";
